+++ river/tests/hooking/test__transition_hooking.py
@@ -1,5 +1,5 @@
 from django.contrib.contenttypes.models import ContentType
-from hamcrest import equal_to, assert_that, has_entry, none, all_of, has_key
+from hamcrest import equal_to, assert_that, has_entry, none, all_of, has_key, has_length, is_not
 
 from river.models import TransitionApproval
 from river.models.factories import PermissionObjectFactory, UserObjectFactory, StateObjectFactory, WorkflowFactory, TransitionApprovalMetaFactory
@@ -56,10 +56,11 @@ class TransitionHooking(BaseHookingTest):
         last_approval = TransitionApproval.objects.get(object_id=workflow_object.model.pk, source_state=state2, destination_state=state3)
 
         output = self.get_output()
-        assert_that(output, has_key("hook"))
-        assert_that(output["hook"], has_entry("type", "on-transit"))
-        assert_that(output["hook"], has_entry("when", AFTER))
-        assert_that(output["hook"], has_entry(
+        assert_that(output, has_length(1))
+        assert_that(output[0], has_key("hook"))
+        assert_that(output[0]["hook"], has_entry("type", "on-transit"))
+        assert_that(output[0]["hook"], has_entry("when", AFTER))
+        assert_that(output[0]["hook"], has_entry(
             "payload",
             all_of(
                 has_entry(equal_to("workflow_object"), equal_to(workflow_object.model)),
@@ -111,10 +112,11 @@ class TransitionHooking(BaseHookingTest):
 
         last_approval = TransitionApproval.objects.get(object_id=workflow_object.model.pk, source_state=state2, destination_state=state3)
         output = self.get_output()
-        assert_that(output, has_key("hook"))
-        assert_that(output["hook"], has_entry("type", "on-transit"))
-        assert_that(output["hook"], has_entry("when", AFTER))
-        assert_that(output["hook"], has_entry(
+        assert_that(output, has_length(1))
+        assert_that(output[0], has_key("hook"))
+        assert_that(output[0]["hook"], has_entry("type", "on-transit"))
+        assert_that(output[0]["hook"], has_entry("when", AFTER))
+        assert_that(output[0]["hook"], has_entry(
             "payload",
             all_of(
                 has_entry(equal_to("workflow_object"), equal_to(workflow_object.model)),
@@ -122,3 +124,56 @@ class TransitionHooking(BaseHookingTest):
 
             )
         ))
+
+    def test_shouldInvokeCallbackForTheOnlyGivenIteration(self):
+        authorized_permission = PermissionObjectFactory()
+        authorized_user = UserObjectFactory(user_permissions=[authorized_permission])
+
+        state1 = StateObjectFactory(label="state1")
+        state2 = StateObjectFactory(label="state2")
+        state3 = StateObjectFactory(label="state3")
+
+        content_type = ContentType.objects.get_for_model(BasicTestModel)
+        workflow = WorkflowFactory(initial_state=state1, content_type=content_type, field_name="my_field")
+        meta1 = TransitionApprovalMetaFactory.create(
+            workflow=workflow,
+            source_state=state1,
+            destination_state=state2,
+            priority=0,
+            permissions=[authorized_permission]
+        )
+
+        TransitionApprovalMetaFactory.create(
+            workflow=workflow,
+            source_state=state2,
+            destination_state=state3,
+            priority=0,
+            permissions=[authorized_permission]
+        )
+
+        TransitionApprovalMetaFactory.create(
+            workflow=workflow,
+            source_state=state3,
+            destination_state=state1,
+            priority=0,
+            permissions=[authorized_permission]
+        )
+
+        workflow_object = BasicTestModelObjectFactory()
+        workflow_object.model.river.my_field.approve(as_user=authorized_user)
+        workflow_object.model.river.my_field.approve(as_user=authorized_user)
+        workflow_object.model.river.my_field.approve(as_user=authorized_user)
+
+        assert_that(TransitionApproval.objects.filter(meta=meta1), has_length(2))
+        first_approval = TransitionApproval.objects.filter(meta=meta1, iteration=0).first()
+        assert_that(first_approval, is_not(none()))
+
+        self.hook_pre_transition(workflow, state1, state2, iteration=0)
+
+        output = self.get_output()
+        assert_that(output, none())
+
+        workflow_object.model.river.my_field.approve(as_user=authorized_user)
+
+        output = self.get_output()
+        assert_that(output, none())